RPT-22: Refactor UI Source File Structure
